fix: guard openat2 debug-file resolution behind a linux build tag
